마인드맵 노드 디자인 오류 수정
