+++ CapStone_vue/src/components/MindMap/MindMap.vue
@@ -390,6 +390,7 @@ import {
 } from "@/api/projectApi";
 import bestIdeaApi from "@/api/bestIdeaApi"; // 파일 상단에 import 추가
 import { DotLottieVue } from "@lottiefiles/dotlottie-vue";
+import { applyBranchColors, darkenColor } from "../utils/colorUtils";
 
 export default {
   components: {
@@ -2176,6 +2177,10 @@ export default {
         branchColor: selectedNode.value.branchColor, // ✅ 부모 노드의 색상 상속
       });
 
+      // 즉시 색상 적용
+      applyBranchColors(myDiagram);
+      myDiagram.updateAllTargetBindings();
+
       myDiagram.commitTransaction("add suggested node");
     };
 
@@ -2429,88 +2434,6 @@ export default {
         }
       }
     };
-
-    const BRANCH_COLORS = ["#F4A261", "#E76F51", "#D4A373", "#C77966"];
-
-    // hex 색상 조금 더 진하게 만들기
-    function darkenColor(hex, percent = 20) {
-      if (!hex) return "#000000";
-      let num = parseInt(hex.slice(1), 16);
-      let r = (num >> 16) & 0xff;
-      let g = (num >> 8) & 0xff;
-      let b = num & 0xff;
-
-      r = Math.max(0, r - (r * percent) / 100);
-      g = Math.max(0, g - (g * percent) / 100);
-      b = Math.max(0, b - (b * percent) / 100);
-
-      return (
-        "#" +
-        ((1 << 24) | (r << 16) | (g << 8) | b)
-          .toString(16)
-          .slice(1)
-          .toUpperCase()
-      );
-    }
-
-    // 모델 로드 후/변경 후마다 호출
-    // 기존 applyBranchColors 함수를 이렇게 수정하세요
-    function applyBranchColors(diagram) {
-      const m = diagram.model;
-      const nodes = m.nodeDataArray;
-      const root = nodes.find((n) => n.parent === 0);
-      if (!root) return;
-
-      diagram.startTransaction("apply branch colors");
-
-      const BRANCH_COLORS = [
-        "#7ED0D6", // light mint blue
-        "#2442A5", // deep blue
-        "#2CAAE1", // cyan-blue
-        "#322B80", // navy purple-blue
-      ];
-
-      // 루트 depth 0
-      m.setDataProperty(root, "depth", 0);
-
-      // 1레벨: 색 배정 + depth 1
-      const firstChildren = nodes.filter((n) => n.parent === root.key);
-      firstChildren.forEach((n, i) => {
-        m.setDataProperty(
-          n,
-          "branchColor",
-          BRANCH_COLORS[i % BRANCH_COLORS.length]
-        );
-        m.setDataProperty(n, "depth", 1);
-        paint(n.key, n.branchColor, 1);
-      });
-
-      // 하위: 부모색 상속 + depth 누적
-      function paint(parentKey, color, depth) {
-        nodes
-          .filter((n) => n.parent === parentKey)
-          .forEach((ch) => {
-            m.setDataProperty(ch, "branchColor", color);
-            m.setDataProperty(ch, "depth", depth + 1);
-            paint(ch.key, color, depth + 1);
-          });
-      }
-
-      // ✅ 링크에도 색상 설정 추가
-      diagram.links.each((link) => {
-        const toNode = link.toNode;
-        if (toNode && toNode.data.branchColor) {
-          m.setDataProperty(link.data, "branchColor", toNode.data.branchColor);
-        }
-        // AI 추천 링크의 경우 isSuggested 속성도 설정
-        if (toNode && toNode.data.isSuggested) {
-          m.setDataProperty(link.data, "isSuggested", true);
-        }
-      });
-
-      diagram.commitTransaction("apply branch colors");
-      diagram.updateAllTargetBindings(); // 바인딩 새로고침
-    }
 
     const initDiagram = () => {
       console.log("📦 initDiagram() 진입");
@@ -3056,13 +2979,13 @@ export default {
             },
             new go.Binding("visible", "isSelected").ofObject(),
 
-            // ✅ 테두리 색
+            // 기존 코드
             new go.Binding("stroke", "isSelected", (sel, obj) => {
               if (!sel) return "transparent";
               const d = obj.part.data;
               const isRoot =
                 d && (d.depth === 0 || d.parent === 0 || d.parent == null);
-              if (isRoot) return "#3B82F6"; // 루트: 검정 테두리
+              if (isRoot) return "#3B82F6";
               if (d.depth === 1)
                 return darkenColor(d.branchColor || "#3B82F6", 25);
               return darkenColor(d.branchColor || "#3B82F6", 25);
